pyproject.toml
[tool.poetry]
name = "griptape-core"
version = "0.1.0"
description = "LLM tools runtime framework."
authors = []
license = "Apache 2.0"
readme = "README.md"
repository = "https://github.com/griptape-ai/griptape-core"

[tool.poetry.dependencies]
python = "^3.9"
attrs = ">=22"
jinja2 = ">=3.1"
docker = ">=6"
stringcase = ">=1"
schema = ">=0.7"
pyyaml = ">=6"
openai = ">=0.27"
langchain = ">=0.0.120"
fastapi = ">=0.90"
uvicorn = ">= 0.20"

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-mock = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
